cleaning up version and building stuff. FOR RUNNING (i.e. matsMethods etc.) on dev we set the process.env.NODE_ENV to "development" via the METEOR_SETTINGS strings ("METEOR_SETTINGS.private.process.NODE_ENV":"development") in the ssl.conf.
on int we set METEOR_SETTINGS.private.process.NODE_ENV":"integration.

on production we leave it out of METEOR_SETTINGS so that it defaults to production.

On a laptop or other development environment you can set it to whatever but it will default to "development".

For the build script we still set the "-e dev|int " flag (no default allowed). The build script will assume development for "-e dev"  and integration for "-e int"

development build machines may not be able to access the appProductionStatus mongo database. Still need to fix that.