Update .dockerignore to exclude .npm directories
MATS ignores the .npm directory in MATScommon as well. This prevents
Meteor from using the .npm/packages/npm-shrinkwrap.json file and
so Meteor pulls in the correct NPM package for the architecture.
Otherwise, Meteor will try to rebuild from source and greatly increase
the image size.

Additionally, reorganized the .dockerignore and updated the exclude
rules to better target artifacts that shouldn't be brought into the
image.